+++ internal/components/datatree/fmt.go
@@ -1,0 +1,7 @@
+package datatree
+
+import "strings"
+
+func trimNewline(s string) string {
+	return strings.TrimSuffix(s, "\n")
+}
