+++ internal/components/datatree/model.go
@@ -1,9 +1,7 @@
 package datatree
 
 import (
-	"fmt"
 	"reflect"
-	"sort"
 	"strings"
 
 	tea "github.com/charmbracelet/bubbletea"
@@ -18,15 +16,15 @@ var (
 )
 
 type Model struct {
-	data      interface{}
-	indentStr string
-	showZero  bool
+	data       interface{}
+	indentSize int
+	showZero   bool
 }
 
 func New(data interface{}) Model {
 	model := Model{
-		data:      data,
-		indentStr: "  ",
+		data:       data,
+		indentSize: 2,
 	}
 
 	return model
@@ -38,107 +36,6 @@ func (m Model) Init() tea.Cmd {
 
 func (m Model) Update(msg tea.Msg) (Model, tea.Cmd) {
 	return m, nil
-}
-
-func trimNewline(s string) string {
-	return strings.TrimSuffix(s, "\n")
-}
-
-func (m Model) renderDataNodeArray(data reflect.Value, indentLevel int) string {
-	result := strings.Builder{}
-
-	elemType := data.Type().Elem()
-
-	for elemType.Kind() == reflect.Ptr {
-		elemType = elemType.Elem()
-	}
-
-	switch elemType.Kind() {
-	case reflect.Struct, reflect.Array, reflect.Slice:
-		result.WriteString("\n")
-		for i := 0; i < data.Len(); i++ {
-			style := lipgloss.NewStyle().
-				Border(lipgloss.RoundedBorder()).
-				MarginLeft((indentLevel + 1) * len(m.indentStr)).
-				PaddingLeft(1).
-				PaddingRight(1)
-			entryStr := m.renderDataNode(data.Index(i), 0)
-			result.WriteString(style.Render(trimNewline(entryStr)))
-			result.WriteString("\n")
-		}
-
-	default:
-		result.WriteString(fmt.Sprintf("%v", data))
-	}
-
-	return result.String()
-}
-
-func (m Model) renderDataNodeStruct(data reflect.Value, indentLevel int) string {
-	result := strings.Builder{}
-	indent := strings.Repeat(m.indentStr, indentLevel)
-
-	fieldNames := []string{}
-
-	result.WriteString("\n")
-
-	for i := 0; i < data.Type().NumField(); i++ {
-		field := data.Type().Field(i)
-
-		if !field.IsExported() {
-			continue
-		}
-
-		fieldNames = append(fieldNames, field.Name)
-	}
-
-	sort.Strings(fieldNames)
-
-	for _, fieldName := range fieldNames {
-		field := data.FieldByName(fieldName)
-
-		if !m.showZero && field.IsZero() {
-			continue
-		}
-
-		for field.Kind() == reflect.Ptr && !field.IsNil() {
-			field = field.Elem()
-		}
-
-		result.WriteString(indent)
-		result.WriteString(styleFieldKey.Render(fieldName + ":"))
-
-		result.WriteString(m.renderDataNode(field, indentLevel+1))
-		result.WriteString("\n")
-		continue
-	}
-
-	return result.String()
-}
-
-func (m Model) renderDataNode(data reflect.Value, indentLevel int) string {
-	for data.Kind() == reflect.Ptr {
-		if data.IsNil() {
-			return "<nil>"
-		}
-
-		data = data.Elem()
-	}
-
-	var result string
-
-	switch data.Kind() {
-	case reflect.Struct:
-		result = m.renderDataNodeStruct(data, indentLevel)
-
-	case reflect.Array, reflect.Slice:
-		result = m.renderDataNodeArray(data, indentLevel)
-
-	default:
-		result = fmt.Sprintf("%v", data)
-	}
-
-	return trimNewline(result)
 }
 
 func (m Model) View() string {
